Update samples for source-faithful metadata

+++ crates/samples/windows/file_dialogs/src/main.rs
@@ -5,7 +5,7 @@ fn main() -> windows_core::Result<()> {
     windows_core::init_mta()?;
 
     unsafe {
-        let dialog: IFileSaveDialog = CoCreateInstance(&FileSaveDialog, None, CLSCTX_ALL as u32)?;
+        let dialog: IFileSaveDialog = CoCreateInstance(&FileSaveDialog, None, CLSCTX_ALL)?;
 
         let filters = [
             COMDLG_FILTERSPEC {
